😜🎰 Checkpoint ./komutlar/botekle-onayla.js:23205691/357

+++ komutlar/botekle-onayla.js
@@ -7,6 +7,7 @@ exports.run = function(client, message, args) {
   let yetkili = message.author
   let sahip = message.guild.members.get(args[0])
 	let botisim = message.guild.members.get(args[1])
+  let botisim2 = args[1]
   let prefix = args[2]
   
 	let log = "758379780751491143" // bot eklendi / onaylandı / reddedildi kanalı
@@ -15,9 +16,10 @@ exports.run = function(client, message, args) {
 
 	if (!botisim) return message.channel.send(`:no_entry: Botun idsini yazmalısın.`).then(msg => msg.delete(10000))
   if(!sahip) return message.reply("Bir sahip id girmen gerekiyor")
+  if(!prefix) return message.reply('Prefixi yazmalısın')
   message.delete()
     sahip.addRole(rol)
-    sahip.setNickname()
+    message.guild.members.get(botisim2).setNickname(` [${prefix}] ` + `${botisim.username}`)
   botisim.addRole(b)
   let embedd = new Discord.RichEmbed()
   .setDescription(` <a:tik4:756946179530424541> | **Tebrikler!** ${botisim} **adlı botun onaylandı.Developer permin verildi** \n\n  🔏 | **Onaylayan yetkili =** ${yetkili} `)
